# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Polar Productos API",
        "description": "API para la gestión de productos, clientes, rutas, precios y descuentos de Polar.",
        "contact": {
            "email": "soporte@polar.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/truncate-customers": {
            "delete": {
                "tags": [
                    "Cargas Masivas - MasterCustomer"
                ],
                "summary": "Truncar tablas de clientes",
                "description": "Vacía todas las tablas relacionadas con datos maestros de cliente (customer_routes, customers, info_types, customer_frequencies, customer_regions, customer_branches, customer_groups) respetando el orden inverso de dependencias.",
                "operationId": "cc1f5ccd8b68972af6bc8dc04c0666fb",
                "responses": {
                    "200": {
                        "description": "Tablas truncadas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Truncate completado: 7854 registro(s) eliminado(s)"
                                        },
                                        "data": {
                                            "properties": {
                                                "customer_routes": {
                                                    "type": "integer",
                                                    "example": 5200
                                                },
                                                "customers": {
                                                    "type": "integer",
                                                    "example": 2100
                                                },
                                                "info_types": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "customer_frequencies": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "customer_regions": {
                                                    "type": "integer",
                                                    "example": 24
                                                },
                                                "customer_branches": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "customer_groups": {
                                                    "type": "integer",
                                                    "example": 8
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Ingresa el token con el prefijo Bearer. Ejemplo: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}